gateway/app/requirements.txt
fastapi==0.128
uvicorn==0.40.0
python-magic==0.4.27
httpx==0.28.0
python-multipart==0.0.9
Pillow==11.2.1
pillow-heif==0.21.0
markitdown[all]
pymupdf
extract-msg
